requirements.txt
yt-dlp==2024.11.18
python-dotenv==1.0.1
pytube==15.0.0
discord-py-interactions[voice]==5.13.1